fix: add not found to delete and findbyid

+++ src/main/java/com/ada/santander/coders/locadora/service/AgenciaService.java
@@ -10,10 +10,7 @@ import com.ada.santander.coders.locadora.response.CepResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.server.ResponseStatusException;
 import reactor.core.publisher.Mono;
 
 import java.util.ArrayList;
@@ -52,7 +49,11 @@ public class AgenciaService {
     }
 
     public Agencia atualizarAgencia(Long id, AgenciaDTO agenciaAtualizado) {
-        Agencia agenciaExistente = this.buscarAgenciaPorId(id);
+        Optional<Agencia> agenciaExistenteOpt = this.buscarAgenciaPorId(id);
+        if(agenciaExistenteOpt.isEmpty()){
+            return null;
+        }
+        Agencia agenciaExistente = agenciaExistenteOpt.get();
         Agencia agenciaNovo = new Agencia();
         agenciaNovo.setTamanhoMaximoDaFrota(agenciaAtualizado.getTamanhoMaximoDaFrota());
 
@@ -65,22 +66,17 @@ public class AgenciaService {
     }
 
     public Agencia deletarAgencia(Long id) {
-        Agencia agenciaExistente = this.buscarAgenciaPorId(id);
+        Optional<Agencia> agenciaExistenteOpt = this.buscarAgenciaPorId(id);
+        if(agenciaExistenteOpt.isEmpty()){
+            return null;
+        }
+        Agencia agenciaExistente =agenciaExistenteOpt.get();
         agenciaRepository.delete(agenciaExistente);
         return agenciaExistente;
     }
 
-    @Transactional
-    public Agencia buscarAgenciaPorId(Long id) {
-        return agenciaRepository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(
-                        HttpStatus.NOT_FOUND, "Agencia com Id " + id
-                ));
-    }
-
-    private boolean verificarAgenciaExistente(Long id) {
-        Optional<Agencia> agenciaExistente = agenciaRepository.findById(id);
-        return agenciaExistente.isPresent();
+    public Optional<Agencia> buscarAgenciaPorId(Long id) {
+        return agenciaRepository.findById(id);
     }
 
     private Mono<CepResponse> consultaCep(String cep) {
@@ -96,21 +92,22 @@ public class AgenciaService {
         return cep != null && cep.matches(CEP_REGEX);
     }
 
-  private Endereco getEnderecoByCep (String cep){
+    private Endereco getEnderecoByCep(String cep) {
 
-      Optional<Endereco> cepExitente = enderecoRepository.findById(cep);
-      if(cepExitente.isPresent()){
-          return cepExitente.get();
-      }else{
-          CepResponse cepResponse = consultaCep(cep).block();
-          Endereco cepNovo = new Endereco();
-          cepNovo.setCep(cepResponse.getCep());
-          cepNovo.setLogradouro(cepResponse.getLogradouro());
-          cepNovo.setBairro(cepResponse.getBairro());
-          cepNovo.setUf(cepResponse.getUf());
-          cepNovo.setCidade(cepResponse.getLocalidade());
-          cepNovo.setRegiao(cepResponse.getRegiao());
-          return enderecoRepository.save(cepNovo);
-      }
-  }
+        Optional<Endereco> cepExitente = enderecoRepository.findById(cep);
+        if (cepExitente.isPresent()) {
+            return cepExitente.get();
+        } else {
+            CepResponse cepResponse = consultaCep(cep).block();
+            Endereco cepNovo = new Endereco();
+            cepNovo.setCep(cepResponse.getCep());
+            cepNovo.setLogradouro(cepResponse.getLogradouro());
+            cepNovo.setBairro(cepResponse.getBairro());
+            cepNovo.setUf(cepResponse.getUf());
+            cepNovo.setCidade(cepResponse.getLocalidade());
+            cepNovo.setRegiao(cepResponse.getRegiao());
+            return enderecoRepository.save(cepNovo);
+        }
+    }
+
 }
